Changes tempdir to be non-local
The current configuration creates a temp dir in the local folder. This
can be problematic because Kotlin takes a long time to download and
install. If you imagine you run the command in a directory where you are
working on another project, this could pose issues for version control
if you need to switch between branches or if you accidentally commit the
tempdir (which normally should be avoided but mistakes are made
sometimes).

This configures the tempdir to be in a global /tmp (which should exist
on Linux and MacOS, the 2 platforms that the script tests for) so that
the user is not blocked.

Include omitted global tempdir

use t flag

the t flag attempts to use the global $TMPDIR variable if it exists,
defaulting to /tmp if it does not